+++ packages/teams-js/test/public/appInitialization.spec.ts
@@ -1,5 +1,5 @@
 import * as AppFile from '../../src/public/app';
-import { appInitialization } from '../../src/public/appInitialization';
+import * as appInitialization from '../../src/public/appInitialization';
 import { version } from '../../src/public/version';
 import { Utils } from '../utils';
