+++ packages/teams-js/test/public/calendar.spec.ts
@@ -2,7 +2,7 @@ import { errorLibraryNotInitialized } from '../../src/internal/constants';
 import { GlobalVars } from '../../src/internal/globalVars';
 import { FrameContexts } from '../../src/public';
 import { app } from '../../src/public/app';
-import { calendar } from '../../src/public/calendar';
+import * as calendar from '../../src/public/calendar';
 import { _minRuntimeConfigToUninitialize } from '../../src/public/runtime';
 import { validateCalendarDeepLinkPrefix } from '../internal/deepLinkUtilities.spec';
 import { Utils } from '../utils';
